Fix applying to a room
WG-gesucht.de changed their form in a way that makes it only usable as a
registered and loggedin user. That's because of this new EU thingy where
you have to accept their privacy policy. This means this application
requires to act on your behalf. You have to set the properties
login.username and login.password with your wg-gesucht.de login
credentials.